refactor: make MockServer reusable.
